Update README.md
This code is intended to run a bookie franchise site where a particular bookie bet code can be converted to several other bookie codes depending on the preferences of our user. However, this convertion is made possible with the use of electronic points which will be purchased in the site. The code however is not complete as to meet the site's standard and as such this project is open to contributions.